requirements.txt
requests>=2.31.0
selenium>=4.15.0
openai-whisper>=20231117